+++ CHANGELOG.md
@@ -1,6 +1,18 @@
+## 0.4.1 - Alpha
+* Widget child builder function added
+
+## 0.4.0 - Alpha
+* Added `willHaveDefaultShortsControllers` option to `ShortsPage`
+* Fix `initialIndex` not working as expected
+* Page controller added to builder options for more user controll
+
+## 0.3.0 - Alpha
+* Initial index assert guideline
+* Export `MediaKit` so developer is able to call `MediaKit.ensureInitialized()`
+
 ## 0.2.0 - Alpha
 * Initial index option added for `VideosSourceController.fromUrlList`` constructor
 * VideoControllerConfiguration of [media_kit](https://pub.dev/packages/media_kit) exported to final user
 
 ## 0.1.0 - Alpha
-* Initial release.
+* Initial release